+++ engine/src/test/java/net/jqwik/engine/properties/shrinking/ShrinkableBigIntegerTests.java
@@ -6,6 +6,7 @@ import java.util.function.*;
 
 import net.jqwik.api.*;
 import net.jqwik.engine.properties.*;
+import net.jqwik.engine.properties.arbitraries.randomized.*;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.Mockito.*;
@@ -30,8 +31,8 @@ class ShrinkableBigIntegerTests {
 
 	@Example
 	void cannotCreateValueOutsideRange() {
-		assertThatThrownBy( //
-			() -> createShrinkableBigInteger(25, Range.of(50L, 100L))) //
+		assertThatThrownBy(
+			() -> createShrinkableBigInteger(25, Range.of(50L, 100L)))
 			.isInstanceOf(JqwikException.class);
 	}
 
@@ -66,9 +67,10 @@ class ShrinkableBigIntegerTests {
 
 	@Example
 	void shrinkingDistanceOutsideLongRange() {
-		Range<BigInteger> bigIntegerRange = Range.of( //
-			new BigInteger("-1000000000000000000000"), //
-			new BigInteger("1000000000000000000000"));
+		Range<BigInteger> bigIntegerRange = Range.of(
+			new BigInteger("-1000000000000000000000"),
+			new BigInteger("1000000000000000000000")
+		);
 
 		assertThat(
 			new ShrinkableBigInteger(new BigInteger("99999999999999999999"), bigIntegerRange, BigInteger.ZERO).distance())
@@ -84,9 +86,8 @@ class ShrinkableBigIntegerTests {
 	void reportFalsified() {
 		Shrinkable<BigInteger> shrinkable = createShrinkableBigInteger(30, Range.of(-100L, 100L));
 
-		ShrinkingSequence<BigInteger> sequence = shrinkable.shrink((TestingFalsifier<BigInteger>) aBigInteger -> aBigInteger
-																													 .compareTo(BigInteger
-																																	.valueOf(10)) < 0);
+		ShrinkingSequence<BigInteger> sequence =
+			shrinkable.shrink((TestingFalsifier<BigInteger>) aBigInteger -> aBigInteger.compareTo(BigInteger.valueOf(10)) < 0);
 
 		assertThat(sequence.next(count, reporter)).isTrue();
 		assertThat(sequence.current().value()).isEqualTo(BigInteger.valueOf(13));
@@ -100,7 +101,6 @@ class ShrinkableBigIntegerTests {
 		verifyNoMoreInteractions(valueReporter);
 	}
 
-
 	@Group
 	class Shrinking {
 
@@ -108,11 +108,10 @@ class ShrinkableBigIntegerTests {
 		void downAllTheWay() {
 			Shrinkable<BigInteger> shrinkable = createShrinkableBigInteger(100000, Range.of(5L, 500000L));
 
-			ShrinkingSequence<BigInteger> sequence = shrinkable.shrink((TestingFalsifier<BigInteger>) aBigInteger -> aBigInteger
-																														 .compareTo(BigInteger
-																																		.valueOf(1000)) <= 0);
+			ShrinkingSequence<BigInteger> sequence =
+				shrinkable.shrink((TestingFalsifier<BigInteger>) aBigInteger -> aBigInteger.compareTo(BigInteger.valueOf(1000)) <= 0);
 
-			while (sequence.next(count, reporter));
+			while (sequence.next(count, reporter)) ;
 
 			assertThat(sequence.current().value()).isEqualTo(BigInteger.valueOf(1001));
 			assertThat(counter.get()).isEqualTo(7);
@@ -127,7 +126,7 @@ class ShrinkableBigIntegerTests {
 
 			ShrinkingSequence<BigInteger> sequence = shrinkable.shrink(filteredFalsifier);
 
-			while (sequence.next(count, reporter));
+			while (sequence.next(count, reporter)) ;
 
 			assertThat(sequence.current().value()).isEqualTo(100);
 			assertThat(counter.get()).isEqualTo(8);
@@ -137,9 +136,8 @@ class ShrinkableBigIntegerTests {
 		void upToExplicitShrinkingTarget() {
 			Shrinkable<BigInteger> shrinkable = createShrinkableBigInteger(1000, Range.of(5L, 500000L), 5000L);
 
-			ShrinkingSequence<BigInteger> sequence = shrinkable.shrink((TestingFalsifier<BigInteger>) aBigInteger -> aBigInteger
-																														 .compareTo(BigInteger
-																																		.valueOf(5000)) >= 0);
+			ShrinkingSequence<BigInteger> sequence =
+				shrinkable.shrink((TestingFalsifier<BigInteger>) aBigInteger -> aBigInteger.compareTo(BigInteger.valueOf(5000)) >= 0);
 
 			while (sequence.next(count, reporter)) ;
 
@@ -154,7 +152,7 @@ class ShrinkableBigIntegerTests {
 		return new ShrinkableBigInteger(
 			BigInteger.valueOf(number),
 			bigIntegerRange,
-			ShrinkableBigInteger.defaultShrinkingTarget(BigInteger.valueOf(number), bigIntegerRange)
+			RandomIntegralGenerators.defaultShrinkingTarget(BigInteger.valueOf(number), bigIntegerRange)
 		);
 	}